+++ src/test/java/com/github/f4b6a3/uuid/factory/nonstandard/ShortPrefixCombFactoryTest.java
@@ -10,6 +10,8 @@ import com.github.f4b6a3.uuid.factory.nonstandard.ShortPrefixCombFactory;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import java.security.SecureRandom;
+import java.time.Clock;
 import java.util.Arrays;
 import java.util.Random;
 import java.util.UUID;
@@ -17,13 +19,13 @@ import java.util.concurrent.ThreadLocalRandom;
 
 public class ShortPrefixCombFactoryTest extends UuidFactoryTest {
 
-	private static final long ONE_MINUTE = 60_000;
+	private static final long DEFAULT_INTERVAL = 60_000;
 
 	@Test
 	public void testGetShortPrefixComb() {
 
 		UUID[] list = new UUID[DEFAULT_LOOP_MAX];
-		ShortPrefixCombFactory factory = new ShortPrefixCombFactory();
+		ShortPrefixCombFactory factory = new ShortPrefixCombFactory((Random) null, null);
 
 		for (int i = 0; i < DEFAULT_LOOP_MAX; i++) {
 			list[i] = factory.create();
@@ -38,13 +40,40 @@ public class ShortPrefixCombFactoryTest extends UuidFactoryTest {
 	public void testGetShortPrefixCombCheckTime() {
 
 		UUID[] list = new UUID[DEFAULT_LOOP_MAX];
-		long startTime = (System.currentTimeMillis() / ONE_MINUTE) & 0x000000000000ffffL;
-
+		long startTime = (System.currentTimeMillis() / DEFAULT_INTERVAL) & 0x000000000000ffffL;
 		for (int i = 0; i < DEFAULT_LOOP_MAX; i++) {
 			list[i] = UuidCreator.getShortPrefixComb();
 		}
+		long endTime = (System.currentTimeMillis() / DEFAULT_INTERVAL) & 0x000000000000ffffL;
 
-		long endTime = (System.currentTimeMillis() / ONE_MINUTE) & 0x000000000000ffffL;
+		checkNotNull(list);
+		checkOrdering(list);
+		checkUniqueness(list);
+
+		long previous = 0;
+		for (int i = 0; i < DEFAULT_LOOP_MAX; i++) {
+			long creationTime = list[i].getMostSignificantBits() >>> 48;
+			assertTrue("Comb Guid creation time before start time", startTime <= creationTime);
+			assertTrue("Comb Guid creation time after end time", creationTime <= endTime);
+			assertTrue("Comb Guid sequence is not sorted " + previous + " " + creationTime, previous <= creationTime);
+			previous = creationTime;
+		}
+	}
+
+	@Test
+	public void testGetShortPrefixCombCheckTimeWithDifferentInterval() {
+
+		Random random = new SecureRandom();
+		Clock clock = Clock.systemUTC();
+		int interval = 1000; // increment the prefix every 1 second interval
+		ShortPrefixCombFactory factory = new ShortPrefixCombFactory(random, clock, interval);
+
+		UUID[] list = new UUID[DEFAULT_LOOP_MAX];
+		long startTime = (System.currentTimeMillis() / interval) & 0x000000000000ffffL;
+		for (int i = 0; i < DEFAULT_LOOP_MAX; i++) {
+			list[i] = factory.create();
+		}
+		long endTime = (System.currentTimeMillis() / interval) & 0x000000000000ffffL;
 
 		checkNotNull(list);
 		checkOrdering(list);