+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 Nothing unreleased.
 
+## [4.6.1] - 2022-03-19
+
+Added time interval parameter for `ShortPrefixCombFactory` and `ShortSuffixCombFactory` constructors. #64
+
 ## [4.6.0] - 2022-03-12
 
 Added `UuidComparator` for comparing and sorting UUIDs. #62
@@ -20,7 +24,7 @@ Remove the `Base32CrfCodec` created 16 days ago to wait for the new base-32 form
 
 ## [4.4.1] - 2022-02-05
 
-Add `Clock` parameter for COMB factories and time functions for tests.
+Add `Clock` parameter for COMB factories and time functions for tests. #60
 
 ## [4.4.0] - 2022-01-28
 
@@ -318,7 +322,8 @@ Remove `Fingerprint` // too complex
 -   Removed UuidSettings.getNodeIdentifierDeprecated // remove deprecated
 -   Renamed UuidSettings to UuidCreatorSettings
 
-[unreleased]: https://github.com/f4b6a3/uuid-creator/compare/uuid-creator-4.6.0...HEAD
+[unreleased]: https://github.com/f4b6a3/uuid-creator/compare/uuid-creator-4.6.1...HEAD
+[4.6.1]: https://github.com/f4b6a3/uuid-creator/compare/uuid-creator-4.6.0...uuid-creator-4.6.1
 [4.6.0]: https://github.com/f4b6a3/uuid-creator/compare/uuid-creator-4.5.0...uuid-creator-4.6.0
 [4.5.0]: https://github.com/f4b6a3/uuid-creator/compare/uuid-creator-4.4.1...uuid-creator-4.5.0
 [4.4.1]: https://github.com/f4b6a3/uuid-creator/compare/uuid-creator-4.4.0...uuid-creator-4.4.1